Refactored common code for files and tasks controller
