+++ app/controllers/organizations/adopter_fosterer/adopted_pets/files_controller.rb
@@ -1,12 +1,7 @@
 module Organizations
   module AdopterFosterer
     module AdoptedPets
-      class FilesController < Organizations::BaseController
-        skip_verify_authorized only: %i[index]
-
-        def index
-          @pet = Pet.find(params[:adopted_pet_id])
-        end
+      class FilesController < Organizations::AdopterFosterer::FilesBaseController
       end
     end
   end
