+++ app/controllers/organizations/adopter_fosterer/adopted_pets/tasks_controller.rb
@@ -1,19 +1,7 @@
 module Organizations
   module AdopterFosterer
     module AdoptedPets
-      class TasksController < Organizations::BaseController
-        before_action :context_authorize!
-
-        def index
-          @pet = Pet.find(params[:adopted_pet_id])
-          @tasks = @pet.tasks.is_not_completed
-        end
-
-        private
-
-        def context_authorize!
-          authorize! with: Organizations::AdopterFosterer::TaskPolicy
-        end
+      class TasksController < Organizations::AdopterFosterer::TasksBaseController
       end
     end
   end
