+++ app/controllers/organizations/adopter_fosterer/files_base_controller.rb
@@ -1,0 +1,11 @@
+module Organizations
+  module AdopterFosterer
+    class FilesBaseController < Organizations::BaseController
+      skip_verify_authorized only: %i[index]
+
+      def index
+        @pet = Pet.find(params[:pet_id])
+      end
+    end
+  end
+end
